pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "src"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Your name (or your organization/company/team)" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "datasets>=3.5.1",
    "dvc>=3.59.2",
    "evaluate>=0.4.3",
    "loguru",
    "numpy>=2.2.5",
    "pip",
    "polars>=1.29.0",
    "python-dotenv",
    "ruff",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
    "tqdm",
    "transformers[torch]>=4.51.3",
    "typer",
]
requires-python = ">3.10"


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    #Pylint
    "PL",
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true
